Create Concordance skeleton
Concordance class is created and given basic functionality.
Implementation of the Concordance was done with a map of word->occurences,
but also the pimpl method was used to encapsulate this implementation since
it might change in the future (i.e. use of an unordered map instead)

Concordances can now only be created from Concordance::makeFromSentences.
Concordance::makeFromFiles also exist but it is not yet implemented: it
only exists to allow a test-driven development. Simple unit tests are also
added which currently fail. Thus these tests were set to be disabled.

File main.cpp contains all the application related code that will
create a concordance. It is responsible for deciding the running mode
according to the command line args given. Application will run in help
mode if -h or --help is given.

Finally, some Formatting functions were defined in a seperate file which
will be used to print the concordance in either console or file