Introduce TextDocumentTraveller
A TextDocumentTraveller is an object which travels a plain text document in chunks.
A chunk is considered to be any series of characters till they reach a whitespace.

TextDocumentTravellers offers also the capability to put back an extracted word.
This will be needed for edge cases of concordance creation

Travelling of document and parsing into chunks is not done with getLine
method to avoid cases of text documents where new lines are not present.
Each time an interaction with the text stream is required, a series of
chunks is parsed and buffered for a slight boost of performance