Sanitize words before adding to concordance
A concordance consists of words in lowercase. During travelling of the text document, found words can be in uppercase, or contain symbols like commas, exclamation marks etc.

A sanitization is needed to
-> convert uppercase characters to lowercase
-> remove any symbols of the word

To achieve this sanitization, two different stratgies were implemented, one for the normal words and one for abbreviations. Since the WordSanitizer can be created multiple times, in the parsing process, these strategies were created as singleton to avoid performance overhead