Finalize parsing of a Text Document
In this changeset the parsing of TextDocument is altered and finalized. In the previous implementation, the TextDocumentTraveller parsed the document into chunks on white space delimeters.
This was just a first implementation to go forward. Main drawback was that it created extra overhead on clients of TextDocumentTraveller, since they would have to reparse the chunk for words.

In this changeset, the whole parsing logic was reconsidered. Travelling of the document will be done in an 'iterator-like' way. Each parsed element can be:
-> a Word, like 'Angelo' or b2b, or even an edge case of b@11$h*t.
-> a symbol, which are not alphanumeric characters that are considered to terminate a word (i.e. .?:,"'<>[]{}() etc)

Instead of using a hierarchy link between a word and symbol, I chose to implement it with std::variant of C++, mainly because these two classes do not have or need a common interface.

Building of the actual concordance is done with a visitor, which according with the current element type decides if it will add it in the concordance or just change the sentence