Fix assertion triggered from wiki dataset
A dataset of wiki I used for testing contained special characters
which triggered an assertion on std::isalnum function

The header <ctype.h> declares several functions useful for classifying
and mapping characters. In all cases the argument is an int, the value of
which shall be representable as an unsigned char or shall equal the
value of the macro EOF. If the argument has any other value, the behavior
is undefined.

The cast to unsigned char ensures calling isalpha() and isalnum does not
invoke Undefined Behaviour.