Lighten WordSanitizer
At my first implementation, WordSanitizer was used for
-> Conversion to lowercase
-> Extraction of symbols

This is in contrast to the principal of Single Responsibility. After the changes of changeset 150699b2f3875020438650572e0076f45ce9bae3 WordSanitizer does not need the burden of SymbolExtraction, since no symbols will exist inside a validated word. WordValidation will make sure that an english word (without symbols) will be valid for insertion in the concordance