Update parsing of DocumentElements to handle edge cases
The previous implementation of parsing of DocumentElement followed a
simplified logic on word parsing. When a word terminating character
was found like ? or !, the word was always terminating. This should not
be the case for words like b@!!sh1t, where here the word should continue
and the explanation mark should not be considered as a Symbol, but as a
part of the word.

To do so, the previous function-based parse algorithm was switched to
a class based algorithm which checks character-by-character and decides
to either consume it or mark it as the word end. Main advantage of this
method is that it keeps state of the previous characters, which is necessary
for checks of abbreviations or special character words.