Fix failure of unit testing
This is essentially a backout of changet 6ce8be14c55ed93413d8635d0a9319c5a02a7895
The above changeset although correct was introduced without updating the unit tests it concered

Problem was that when an index, a word or an occurence was made printable,
another space was added to guard it from collapsing with the next contributor
of a Concordance line.

This should not be the responsibility of makePrintable to add such a space.
Instead, a joining function was introduced to handle such a case along with it's unit tests